core: introduce NlaType
In accordance with netlinklib's goal of being as
1) fast
2) flexible/extensible
3) minimalist
as possible, this commit introduces the NlaType class family as a way of allowing users to easily create their own customized netlink message parsers and constructors. NlaType and its subclasses are composed by the user into structures resembling netlink messages. These nested structures expose two primary methods: `to_bytes` and `parse`, which generate specified netlink message or parse specified netlink message respectively.

If a NlaType structure is to be used for serializing, each object in the nested structure should be initialized with a proper value, which will be rendered into a byte-format by the `to_bytes` method.

If used for parsing, the structure defines how parsing should occur. Only attributes included in the structure are parsed; any additional information in the incoming netlink message is ignored for speed. If an object in the NlaType structure is initialized with a specific value, this value is used for filtering messages during parsing: during a netlink dump, if the parsed value of an NlaType with user-provided value does not match this value, parsing of message is aborted and the next message in the dump is checked. These behaviors are implemented in such a way to avoid as much work as possible during parsing, and consequently improve performance.

Going forward, the external API is to be simplified: users will be expected to construct their own parsers and serializers using the provided classes i.e. anything residing in `api_x.py` or `parser_x.py` will be in the domain of the user. These modules are deprecated and will be removed. For an example of how NlaType should be used, see `examples` directory.

Signed-off-by: Michael Dickensheets <michael.dickensheets@ionos.com>