core: add more generic parser filtering mechanism
Prior to this change, filtering was limited to matching user-provided values to parsed ones in order to determine whether or not to bail out of parsing current message.

This adds `filter` argument to scalar NlaTypes and `filters` to NllMsg types, which receive callables to make this determination. The callables received the parsed value and return a boolean: true if result is valid/parsing should continue, false if parsing should be aborted.

Signed-off-by: Michael Dickensheets <michael.dickensheets@ionos.com>