core: remove keyword arguments from NlaTypes
"setters" have been subsumed into general callback framework, and both callbacks and "vals" are now passed as non-keyword arguments to NlaAttr types and are separated there.

The point of this is to make the library syntax less verbose.

Analogous changes were made to NllMsg types, where instead of magic "setters" keyword, callbacks are given using the kwarg matching the field.

Signed-off-by: Michael Dickensheets <michael.dickensheets@ionos.com>